cleanup github module
